Protocol Downscaler wired up
