+++ src/api/Service/Inc/MidiClientManager.h
@@ -63,6 +63,7 @@ private:
                         _In_ handle_t,
                         _In_ MidiFlow,
                         _In_ wil::com_ptr_nothrow<CMidiPipe>&,
+                        _In_ wil::com_ptr_nothrow<CMidiPipe>&,
                         _In_ wil::com_ptr_nothrow<CMidiPipe>&);
 
     HRESULT
